+++ libfabric.vcxproj
@@ -575,7 +575,7 @@
     <ClInclude Include="include\ofi_iov.h" />
     <ClInclude Include="include\ofi_indexer.h" />
     <ClInclude Include="include\fi_list.h" />
-    <ClInclude Include="include\fi_lock.h" />
+    <ClInclude Include="include\ofi_lock.h" />
     <ClInclude Include="include\fi_mem.h" />
     <ClInclude Include="include\fi_osd.h" />
     <ClInclude Include="include\ofi_perf.h" />
