+++ prov/gni/include/gnix_smrn.h
@@ -34,7 +34,7 @@
 #define PROV_GNI_INCLUDE_GNIX_SMRN_H
 
 #include "include/fi_list.h"
-#include "include/fi_lock.h"
+#include "include/ofi_lock.h"
 #include "config.h"
 
 #include "gnix_mr_notifier.h"
