+++ prov/mlx/src/mlx.h
@@ -57,7 +57,7 @@ extern "C" {
 #include <rdma/fi_errno.h>
 #include <rdma/providers/fi_log.h>
 #include <fi.h>
-#include <fi_lock.h>
+#include <ofi_lock.h>
 #include <fi_list.h>
 #include "fi_enosys.h"
 #include <fi_mem.h>
